alias to be server specific

+++ commands/technical/alias/del.js
@@ -5,9 +5,9 @@ const cmd = new command_1.command({
     name: 'del',
     run: async (client, msg, args) => {
         const alias = args.shift();
-        if (client.aliases.has(alias)) {
-            const aliasfcmd = client.aliases.get(alias);
-            const aliasdb = client.db.db('aliases', {});
+        if (client.aliases.has(msg.guild.id) && client.aliases.get(msg.guild.id).has(alias)) {
+            const aliasfcmd = client.aliases.get(msg.guild.id).get(alias);
+            const aliasdb = client.db.db(msg.guild.id, { traversepath: ['aliases'] });
             if (aliasdb.delete({ alias: alias, fcmd: aliasfcmd })) {
                 msg.reply(`Successfully deleted ${alias}.`);
                 client.aliases.delete(alias);
